Update ReceiveMode to use History directly, and now all GUI tests pass
